Refactored project into separate files (preparing for tournament functionality)
